+++ libft/ft_countwords.c
@@ -6,7 +6,7 @@
 /*   By: ratroncy <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/13 16:17:53 by ratroncy          #+#    #+#             */
-/*   Updated: 2017/11/13 17:19:45 by ratroncy         ###   ########.fr       */
+/*   Updated: 2017/11/20 11:05:21 by ratroncy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_countwords(const char *s, int c)
 {
-	int		i;
-	int		words;
+	size_t	i;
+	size_t	words;
 
 	i = 0;
 	words = 0;
