+++ libft/ft_strncat.c
@@ -6,16 +6,16 @@
 /*   By: ratroncy <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/15 18:19:41 by ratroncy          #+#    #+#             */
-/*   Updated: 2017/11/14 18:11:25 by ratroncy         ###   ########.fr       */
+/*   Updated: 2017/11/20 11:13:18 by ratroncy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strncat(char *dest, char *src, int nb)
+char	*ft_strncat(char *dest, char *src, size_t nb)
 {
-	int i;
-	int j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
